[CHORE] - Move adapter classes
