+++ app/src/main/java/com/example/marvelapp/presentation/characters/adapters/CharactersLoadMoreStateAdapter.kt
@@ -1,10 +1,10 @@
-package com.example.marvelapp.presentation.characters
+package com.example.marvelapp.presentation.characters.adapters
 
 import android.view.ViewGroup
 import androidx.paging.LoadState
 import androidx.paging.LoadStateAdapter
 
-class CharactersLoadingStateAdapter(
+class CharactersLoadMoreStateAdapter(
     private val retry: () -> Unit
 ) : LoadStateAdapter<CharactersLoadMoreStateViewHolder>() {
