+++ app/src/main/java/com/example/marvelapp/presentation/characters/adapters/CharactersLoadMoreStateViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.marvelapp.presentation.characters
+package com.example.marvelapp.presentation.characters.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
